LPS-95015 Move the classes to the com.liferay.site.service.test package.
